refactor: make PanelHeaderButton take a classname property
