+++ app/components/PanelHeaderButton/PanelHeaderButton.scss
@@ -5,7 +5,7 @@
     display: flex;
     margin-right: 10px;
     align-items: center;
-    width: 200px;
+    width: 170px;
     font-size: 1rem;
     align-items: center;
     cursor: pointer;
