+++ app/components/TokenSale/TokenSalePanel/TokenSalePanel.scss
@@ -22,6 +22,10 @@
   justify-content: space-between;
 }
 
+.tokenSaleHeaderButton {
+  width: 200px;
+}
+
 .warningIcon {
   path {
     fill: #ee6d66;
